refactor(agent-event-protocol): extract wire contract into standalone modules
Split index.ts into agent-event.types.ts and wire-content.types.ts so the barrel only re-exports and internal modules avoid importing from it.
