+++ packages/agent-event-protocol/src/wire-content.types.ts
